+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file, following t
 
 ## [Unreleased]
 
+- Fix compression not being applied to mvsj_to_mvsx function, MVSX node, and molstar widget
+
 ## [v1.8.0] - 2025-11-05
 
 - Add support for `dx` volume format